services/greeter/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
